+++ src/deisa/dask/handshake.py
@@ -27,6 +27,7 @@
 # POSSIBILITY OF SUCH DAMAGE.
 # =============================================================================
 import logging
+from typing import Optional
 
 from distributed import Client, Future, get_client, Event
 
@@ -38,9 +39,8 @@ logger = logging.getLogger(__name__)
 class Handshake:
     _DEISA_HANDSHAKE_ACTOR_FUTURE_VARIABLE = 'deisa_handshake_actor_future'
     _DEISA_WAIT_FOR_DONE_EVENT = 'deisa_handshake_done'
-    _DEISA_WAIT_FOR_BRIDGE_READY_EVENT = 'deisa_handshake_bridge_ready'
     _DEISA_WAIT_FOR_BRIDGE_DONE_EVENT = 'deisa_handshake_bridge_done'
-    _DEISA_WAIT_FOR_ANALYTICS_READY_EVENT = 'deisa_handshake_analytics_ready'
+    _DEISA_WAIT_FOR_GO_EVENT = 'deisa_handshake_go'
 
     class HandshakeActor:
         bridges_ready = []
@@ -50,11 +50,20 @@ class Handshake:
 
         def __init__(self):
             logger.debug('HandshakeActor.__init__()')
+            self.analytics_ready = False
             self.bridges_ready = []
             self.bridges_done = []
             self.max_bridges = 0
             self.arrays_metadata = {}
+            self.feedback_queue_size: int = 1024
             self.client = get_client()
+
+        def add_analytics_ready(self, feedback_queue_size: int) -> None | Future:
+            self.analytics_ready = True
+            self.feedback_queue_size = feedback_queue_size
+
+            if self.are_bridges_ready():
+                self.go()
 
         def add_bridge_ready(self, id: int, max: int) -> None:
             if max == 0:
@@ -68,6 +77,9 @@ class Handshake:
 
             self.bridges_ready.append(id)
 
+            if self.is_everyone_ready():
+                self.go()
+
         def add_bridge_done(self, id: int) -> None:
             self.bridges_done.append(id)
             if len(self.bridges_ready) == self.max_bridges:
@@ -79,22 +91,33 @@ class Handshake:
         def get_arrays_metadata(self) -> dict | Future:
             return self.arrays_metadata
 
+        def get_feedback_queue_size(self) -> int | Future:
+            return self.feedback_queue_size
+
         def get_max_bridges(self) -> int | Future:
             return self.max_bridges
 
         def are_bridges_ready(self) -> bool | Future:
             return self.max_bridges != 0 and len(self.bridges_ready) == self.max_bridges
 
-    def __init__(self, who: str, client: Client, **kwargs):
+        def is_everyone_ready(self) -> bool | Future:
+            return self.analytics_ready and self.are_bridges_ready()
+
+        def go(self) -> None:
+            Event(Handshake._DEISA_WAIT_FOR_GO_EVENT, client=self.client).set()
+
+    def __init__(self, who: str, client: Optional[Client], feedback_queue_size=1024, **kwargs):
         self.client = client
         # self.client.direct_to_workers() # TODO
-        self.__handshake_actor = _get_actor(self.client, Handshake.HandshakeActor)
-        assert self.__handshake_actor is not None
+
+        if client:
+            self.__handshake_actor = _get_actor(self.client, Handshake.HandshakeActor)
+            assert self.__handshake_actor is not None
 
         if who == 'bridge':
             self.start_bridge(**kwargs)
         elif who == 'deisa':
-            pass
+            self.start_deisa(feedback_queue_size, **kwargs)
         else:
             raise ValueError("Expecting 'bridge' or 'deisa'.")
 
@@ -103,6 +126,10 @@ class Handshake:
         Bridge must wait for analytics to be ready.
         """
         assert self.__handshake_actor is not None
+
+        if self.client:
+            assert id == 0, "only bridge 0 should have a valid client"
+
         self.__handshake_actor.add_bridge_ready(id, max).result()
 
         # TODO: change this so that the check is not necessarily done on id=0
@@ -110,29 +137,32 @@ class Handshake:
             self.__handshake_actor.set_arrays_metadata(arrays_metadata).result()
         # TODO: check that arrays_metadata is the same for all bridges
 
-        if self.__handshake_actor.are_bridges_ready().result():
-            Event(Handshake._DEISA_WAIT_FOR_BRIDGE_READY_EVENT, client=self.client).set()
+        if kwargs.get('wait_for_go', True):
+            self.wait_for_go()
 
-        if kwargs.get('analytics_ready', False):
-            # bridges wait for analytics to be ready
-            Event(Handshake._DEISA_WAIT_FOR_ANALYTICS_READY_EVENT, client=self.client).wait()
+    def start_deisa(self, feedback_queue_size: int, *args, **kwargs) -> None:
+        self.__handshake_actor.add_analytics_ready(feedback_queue_size).result()
+
+        if kwargs.get('wait_for_go', True):
+            self.wait_for_go()
 
     def get_arrays_metadata(self) -> dict:
         assert self.__handshake_actor is not None
         return self.__handshake_actor.get_arrays_metadata().result()
 
+    def get_feedback_queue_size(self) -> int:
+        assert self.__handshake_actor is not None
+        return self.__handshake_actor.get_feedback_queue_size().result()
+
     def get_nb_bridges(self) -> int:
         assert self.__handshake_actor is not None
         return self.__handshake_actor.get_max_bridges().result()
 
-    def deisa_ready(self):
-        Event(Handshake._DEISA_WAIT_FOR_ANALYTICS_READY_EVENT, client=self.client).set()
-
     def stop_bridge(self, id: int) -> None:
         self.__handshake_actor.add_bridge_done(id).result()
 
-    def wait_for_bridges_to_start(self):
-        Event(Handshake._DEISA_WAIT_FOR_BRIDGE_READY_EVENT, client=self.client).wait()
+    def wait_for_go(self):
+        Event(Handshake._DEISA_WAIT_FOR_GO_EVENT, client=self.client).wait()
 
     def wait_for_bridges_to_finish(self):
         Event(Handshake._DEISA_WAIT_FOR_BRIDGE_DONE_EVENT, client=self.client).wait()